+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # [Next](https://github.com/Mossop/bugzilla-ts/compare/v2.1.0...main)
 
+- Added `createBug` and `updateBug` API methods.
+
 # [2.1.0](https://github.com/Mossop/bugzilla-ts/compare/v2.0.0...v2.1.0)
 
 - Added `getComment`, `getBugComments` and `createComment` API methods.